[366] add matrix logging capability
