+++ src/test/java/nz/cri/gns/NZSHM22/util/DataLoggerTest.java
@@ -1,4 +1,4 @@
-package nz.cri.gns.NZSHM22.opensha.inversion;
+package nz.cri.gns.NZSHM22.util;
 
 import static org.junit.Assert.*;
 
@@ -7,12 +7,15 @@ import java.nio.file.Files;
 import java.util.stream.Collectors;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipFile;
+
+import cern.colt.matrix.tdouble.DoubleMatrix2D;
+import cern.colt.matrix.tdouble.impl.SparseDoubleMatrix2D;
 import org.junit.Test;
 
-public class LoggingCompletionCriteriaTest {
+public class DataLoggerTest {
 
     static String format(double value) {
-        return LoggingCompletionCriteria.MultiZipLog.format(value);
+        return DataLogger.format(value);
     }
 
     @Test
@@ -46,8 +49,8 @@ public class LoggingCompletionCriteriaTest {
 
         File tempDir = Files.createTempDirectory("zipLog").toFile();
 
-        LoggingCompletionCriteria.MultiZipLog log =
-                new LoggingCompletionCriteria.MultiZipLog(tempDir.getAbsolutePath(), "testLog", 30);
+        DataLogger.MultiZipLog log =
+                new DataLogger.MultiZipLog(tempDir.getAbsolutePath(), "testLog", 30);
         log.addHeader(headerFile, "a,b,c\n");
 
         log.nextIndex(0);
@@ -73,5 +76,57 @@ public class LoggingCompletionCriteriaTest {
         assertEquals("4.0,5.0,6.0\n7.0,8.0,9.0", actual);
         actual = getFile(new File(tempDir, "testLog[2-2].zip"), arrayFile + ".csv");
         assertEquals("10.0,11.0,12.0", actual);
+    }
+
+    @Test
+    public void testDumpVector() throws IOException {
+
+        File tempDir = Files.createTempDirectory("dump").toFile();
+        String fileName = tempDir.getAbsolutePath() + "/testDumpVector.csv";
+        double[] sourceValue =
+                new double[] {
+                    0, 1.0, 1.2, 3.1456789, Double.NaN, Double.POSITIVE_INFINITY, Double.MIN_VALUE
+                };
+
+        DataLogger.dump(fileName, sourceValue);
+
+        String actual = "";
+
+        try (BufferedReader reader = new BufferedReader(new FileReader(fileName))) {
+            actual += reader.lines().collect(Collectors.joining("\n"));
+        }
+
+        assertEquals("0,1.0,1.2,3.146,NaN,Infinity,4.9E-324", actual);
+    }
+
+    @Test
+    public void testDumpMatrix() throws IOException {
+
+        File tempDir = Files.createTempDirectory("dump").toFile();
+        String fileName = tempDir.getAbsolutePath() + "/testDumpMatrix.csv";
+
+        DoubleMatrix2D sourceValue = new SparseDoubleMatrix2D(5, 5);
+        sourceValue.set(0, 0, 42);
+        sourceValue.set(0, 1, Double.NEGATIVE_INFINITY);
+        sourceValue.set(0, 2, 2);
+        sourceValue.set(0, 3, 3);
+        sourceValue.set(0, 4, 4);
+        sourceValue.set(4, 4, 13);
+        sourceValue.set(3, 2, 7);
+        DataLogger.dump(fileName, sourceValue);
+
+        String actual = "";
+
+        try (BufferedReader reader = new BufferedReader(new FileReader(fileName))) {
+            actual += reader.lines().collect(Collectors.joining("\n"));
+        }
+
+        assertEquals(
+                "42.0,-Infinity,2.0,3.0,4.0\n"
+                        + "0,0,0,0,0\n"
+                        + "0,0,0,0,0\n"
+                        + "0,0,7.0,0,0\n"
+                        + "0,0,0,0,13.0",
+                actual);
     }
 }
